% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAICBIC.R
\name{computeAIC}
\alias{computeAIC}
\title{Akaike information criterion}
\usage{
computeAIC(log_likelihood, num_parameters)
}
\arguments{
\item{log_likelihood}{log likelihood function value}

\item{num_parameters}{the number of parameters estimated}
}
\value{
the AIC as a number
}
\description{
This function computes Akaike information criterion
}
