% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAICBIC.R
\name{computeBIC}
\alias{computeBIC}
\title{Bayesian information criterion}
\usage{
computeBIC(log_likelihood, num_parameters, num_observations)
}
\arguments{
\item{log_likelihood}{log likelihood function value}

\item{num_parameters}{the number of parameters estimated}

\item{num_observations}{the number of observations used to estimate the parameters}
}
\value{
the BIC as a number
}
\description{
This function computes Bayesian information criterion
}
