% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOneFactor.R
\name{fitOneFactor}
\alias{fitOneFactor}
\title{Fit a One-Factor Copula Model}
\usage{
fitOneFactor(df, copula, params = list())
}
\arguments{
\item{df}{A data frame containing uniform variables with \code{nrow(df)}
observations and \code{ncol(df)} variables.}

\item{copula}{A character string specifying the copula to be fitted.
Supported values are \code{'Gaussian'}, \code{'Clayton'}, \code{'Gumbel'}, and \code{'rGumbel'}.}

\item{params}{A list of parameters for the copula model.
Default is \code{list(n_nodes = 35, init_val = rep(0.5, ncol(df)))} for Gaussian copula,
\code{list(n_nodes = 25, init_val = rep(2, ncol(df)))} for Clayton, Gumbel and reflected Gumbel copulas
.}
}
\value{
A list containing:
\describe{
\item{\code{minimum}}{The estimated minimum of the negative log-likelihood.}
\item{\code{estimate}}{The optimal parameter values.}
\item{\code{gradient}}{The gradient at the estimated minimum.}
}
}
\description{
This function fits a one-factor copula model to the given data frame of uniform variables.
Supported copulas include 'Gaussian', 'Clayton', 'Gumbel', and 'rGumbel'.
}
