% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOneFactorDynamic.R
\name{fitOneFactorDynamic}
\alias{fitOneFactorDynamic}
\title{Fit a One-Factor Copula Model With Dynamic Dependence}
\usage{
fitOneFactorDynamic(df, copula, params = list())
}
\arguments{
\item{df}{A data frame containing uniform variables with \code{nrow(df)}
observations and \code{ncol(df)} variables.}

\item{copula}{A character string specifying the copula to be fitted.
Supported values are \code{'Gaussian'}, \code{'Clayton'}, \code{'Gumbel'}, and \code{'rGumbel'}.}

\item{params}{A list of parameters for the copula model. n_nodes is the number of Gauss-Legendre quadrature points,
init_val is the initial value of the parameter values to begin our search, period is the rolling window size
Default is \code{list(n_nodes = 35, init_val = rep(0.5, ncol(df)), period = 90)} for Gaussian copula,
\code{list(n_nodes = 25, init_val = rep(2, ncol(df)), period = 90)} for Clayton, Gumbel and reflected Gumbel copulas
.}
}
\value{
A list containing:
\describe{
\item{\code{minimum}}{The estimated minimum of the negative log-likelihood.}
\item{\code{estimate}}{The optimal parameter values.}
\item{\code{gradient}}{The gradient at the estimated minimum.}
}
}
\description{
This function fits a one-factor copula model with dynamic dependence to the given data frame of uniform variables.
Supported copulas include 'Gaussian', 'Clayton', 'Gumbel', and 'rGumbel'.
}
