% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTailDependence.R
\name{powerTailDependence}
\alias{powerTailDependence}
\title{Tail weighted measure of dependence}
\usage{
powerTailDependence(df, d, p)
}
\arguments{
\item{df}{A data frame containing uniform variables with \code{nrow(df)}
observations and \code{ncol(df)} variables.}

\item{d}{the power used in the power function}

\item{p}{the truncation level satisfying 0 < p < 0.5
.}
}
\value{
A ncol(df) x ncol(df) matrix containing the tail weighted measure of dependence between the variables
}
\description{
This function computes the tail weighted measure of dependence as defined in Krupskyi, P. (2021). Tail-weighted measures of dependence. Dependence Modeling, 9(1), 92-113. https://doi.org/10.1515/demo-2021-0102
}
