% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOneFactor.R
\name{simulateOneFactor}
\alias{simulateOneFactor}
\title{Simulate from a one factor copula}
\usage{
simulateOneFactor(n, copula, d, theta)
}
\arguments{
\item{n}{number of observations to simulate}

\item{copula}{the copula to simulate from, can be one of Gaussian, Clayton, Gumbel and rGumbel}

\item{d}{number of variables}

\item{theta}{the copula parameters as a vector}
}
\value{
A ncol(df) x ncol(df) matrix containing the tail weighted measure of dependence between the variables
}
\description{
This function simulates from a one factor copula model
}
